# src/docs/openapi.dev.json
{
  "openapi": "3.0.3",
  "info": { "title": "Business API (Dev)", "version": "1.0.0-dev", "description": "Full specification including Dev/Debug endpoints" },
  "servers": [ { "url": "http://localhost:3000" } ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "Authorization", "description": "Use 'Authorization: Token ck_...' for customer API key auth" }
    },
    "schemas": {
      "OkWrapper": { "type": "object", "properties": { "success": { "type": "boolean", "example": true }, "data": { "type": "object" } } },
      "ErrorResponse": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } }, "required": ["code", "message"] } }, "required": ["success", "error"] },
...
      "AdminAuditItem": { "type": "object", "properties": { "id": { "type": "string" }, "endpoint": { "type": "string" }, "method": { "type": "string" }, "statusCode": { "type": "integer" }, "responseTimeMs": { "type": "integer" }, "timestamp": { "type": "string" } } }
    }
  },
  "paths": {
...
    "/api/v1/admin/auth/login": { "post": { "tags": ["Auth","Admin"], "summary": "Admin login (JWT)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } }, "required": ["email","password"] } } } }, "responses": { "200": { "description": "Logged in", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkWrapper" } } } }, "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } } } },
...
    "/api/v1/admin/audit/self": { "get": { "tags": ["Admin"], "summary": "View recent admin actions (audit)", "security": [{"bearerAuth":[]}], "parameters": [ { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } } ], "responses": { "200": { "description": "Audit entries", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAuditItem" } } } } } } }, "401": { "description": "Unauthorized" } } } },
    "/api/v1/admin/dev/reset-password": { "post": { "tags": ["Dev/Debug"], "summary": "DEV ONLY: reset admin password from env", "responses": { "200": { "description": "Reset OK" }, "403": { "description": "Forbidden in prod" } } } },
    "/api/v1/admin/dev/seed": { "post": { "tags": ["Dev/Debug"], "summary": "DEV ONLY: seed admin email/password", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } }, "required": ["email", "password"] } } } }, "responses": { "200": { "description": "Seeded" }, "403": { "description": "Forbidden in prod" } } } },
    "/api/v1/admin/debug/verify-token-detailed": { "post": { "tags": ["Dev/Debug"], "summary": "Verify a plaintext API key against DB", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" } }, "required": ["token"] } } } }, "responses": { "200": { "description": "Verification details" } } } },
    "/api/v1/admin/debug/import-plain-key": { "post": { "tags": ["Dev/Debug"], "summary": "Import a plaintext key from vault to DB", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "plainKey": { "type": "string" } }, "required": ["plainKey"] } } } }, "responses": { "200": { "description": "Imported or already exists" }, "400": { "description": "Not found in vault" } } } },
    "/api/v1/admin/debug/plain-keys": { "get": { "tags": ["Dev/Debug"], "summary": "List recent masked plaintext keys from file vault", "parameters": [ { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } } ], "responses": { "200": { "description": "Keys" } } } },
    "/api/v1/admin/debug/plain-keys-sqlite": { "get": { "tags": ["Dev/Debug"], "summary": "List recent masked keys from sqlite vault", "parameters": [ { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } } ], "responses": { "200": { "description": "Keys" } } } }
  }
}